regexec.c: S_find_byclass(): utf8ness in switch()
This changes this function to have different case: statements in a
switch() depending on whether the target is UTF-8 or not, and the
switch() itself includes the UTF8ness.

This will allow restructuring the code to be cleaner, and to have fewer
conditionals, as many of the case:s first task is to find out the
target's UTF8ness.